Some address are not in the range that used to emulate main memory.
We have to find solutions for these memory.

+++ cm-replay.c
@@ -38,7 +38,7 @@ extern int smp_cpus;
 int cm_run_mode;
 
 /* Use this to get cpu specific info. */
-__thread int cm_coreid;
+__thread uint16_t cm_coreid;
 
 /* Array containing tb execution count for each cpu. */
 uint64_t *cm_tb_exec_cnt;
